src_cpp: Remove move limit (#2)
## Description

Having a move limit makes sense for training, but we don't want to have it for playing against ELF in GTP mode for evaluation. Currently what happens is that if we exceed 722 moves when playing against ELF, ELF will automatically reject the move and end the game.

This PR crudely removes the move limit. 

(Ideally I would set the option `move_cutoff` to the previous value of the move limit, make sure `move_cutoff` is used everywhere, and just specify `--move_cutoff -1` for our evaluations. But I don't think it's worth the time for me to learn the ELF codebase well enough to make that more sophisticated change.)

## Testing

- launch ELF in GTP mode, paste a game of 1000 moves, check that no errors are printed